Changed inventory bag slots server side so they match the client. This is a significant change, so please report any new item issues that crop up.
REQUIRED SQL: required/2014_07_17_inventory.sql
This MUST be run when the server is down.

Increased chance for Sense Heading to level up a bit.

Implemented /shield. The .conf will need to be copied over.

+++ common/eq_packet_structs.h
@@ -2148,7 +2148,7 @@ struct ClickObject_Struct {
 };
 
 struct Shielding_Struct {
-	uint32 target_id;
+	uint16 target_id;
 };
 
 /*
